+++ cursepy/__init__.py
@@ -9,5 +9,5 @@ from cursepy.wrapper import CurseClient, MinecraftWrapper
 
 # Define some metadata here:
 
-__version__ = '2.1.0'
+__version__ = '2.2.0'
 __author__ = 'Owen Cochell'